Move out all commands into their own modules
